requirements.txt
pyqt5
pyserial
cryptography
pyinstaller
